Move PP viewing post assessment feedback
to appointments controller, where we can re-use the SP-focused function.
