+++ server/routes/probationPractitionerReferrals/probationPractitionerReferralsController.test.ts
@@ -241,78 +241,6 @@ describe('GET /probation-practitioner/end-of-service-report/:id', () => {
   })
 })
 
-describe('GET /probation-practitioner/referrals/:id/supplier-assessment/post-assessment-feedback', () => {
-  it('renders a page showing the initial assessment feedback', async () => {
-    const deliusServiceUser = deliusServiceUserFactory.build()
-    const referral = sentReferralFactory.assigned().build()
-    const appointment = initialAssessmentAppointmentFactory.build({
-      appointmentTime: '2021-02-01T13:00:00Z',
-      sessionFeedback: {
-        attendance: {
-          attended: 'yes',
-          additionalAttendanceInformation: 'He was punctual',
-        },
-        behaviour: {
-          behaviourDescription: 'Acceptable',
-        },
-        submitted: false,
-      },
-    })
-    const supplierAssessment = supplierAssessmentFactory.build({
-      appointments: [appointment],
-      currentAppointmentId: appointment.id,
-    })
-    communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
-    interventionsService.getSentReferral.mockResolvedValue(referral)
-    interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
-
-    await request(app)
-      .get(`/probation-practitioner/referrals/${referral.id}/supplier-assessment/post-assessment-feedback`)
-      .expect(200)
-      .expect(res => {
-        expect(res.text).toContain('View feedback')
-        expect(res.text).toContain('Did Alex attend the initial assessment appointment?')
-        expect(res.text).toContain('Yes, they were on time')
-        expect(res.text).toContain('Describe Alex&#39;s behaviour in the assessment appointment')
-        expect(res.text).toContain('Acceptable')
-      })
-  })
-  it('renders an error if there is the referral is not assigned', async () => {
-    const deliusServiceUser = deliusServiceUserFactory.build()
-    const referral = sentReferralFactory.unassigned().build()
-    const supplierAssessment = supplierAssessmentFactory.build({
-      appointments: [],
-    })
-    communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
-    interventionsService.getSentReferral.mockResolvedValue(referral)
-    interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
-
-    await request(app)
-      .get(`/probation-practitioner/referrals/${referral.id}/supplier-assessment/post-assessment-feedback`)
-      .expect(500)
-      .expect(res => {
-        expect(res.text).toContain('Referral has not yet been assigned to a caseworker')
-      })
-  })
-  it('renders an error if there is no current appointment for the initial assessment', async () => {
-    const deliusServiceUser = deliusServiceUserFactory.build()
-    const referral = sentReferralFactory.assigned().build()
-    const supplierAssessment = supplierAssessmentFactory.build({
-      appointments: [],
-    })
-    communityApiService.getServiceUserByCRN.mockResolvedValue(deliusServiceUser)
-    interventionsService.getSentReferral.mockResolvedValue(referral)
-    interventionsService.getSupplierAssessment.mockResolvedValue(supplierAssessment)
-
-    await request(app)
-      .get(`/probation-practitioner/referrals/${referral.id}/supplier-assessment/post-assessment-feedback`)
-      .expect(500)
-      .expect(res => {
-        expect(res.text).toContain('Attempting to view initial assessment feedback without a current appointment')
-      })
-  })
-})
-
 describe('GET /probation-practitioner/referrals/:id/details', () => {
   const intervention = interventionFactory.build()
   const hmppsAuthUser = hmppsAuthUserFactory.build({ firstName: 'John', lastName: 'Smith' })
